removes additional wrappers around image component
