--- packages/terra-core-docs/src/terra-dev-site/doc/image/ChangeLog.3.doc.mdx
@@ -1,7 +1,0 @@
-import { Badge } from 'terra-image/package.json?dev-site-package';
-
-import ChangeLog from 'terra-image/CHANGELOG.md';
-
-<Badge />
-
-<ChangeLog />
