+++ packages/terra-core-docs/src/terra-dev-site/doc/image/ChangeLog.4.doc.mdx
@@ -1,0 +1,7 @@
+import { Badge } from 'terra-image/package.json?dev-site-package';
+
+import ChangeLog from 'terra-image/CHANGELOG.md';
+
+<Badge />
+
+<ChangeLog />
