+++ packages/terra-core-docs/src/terra-dev-site/doc/image/UpgradeGuide.3.doc.mdx
@@ -1,0 +1,12 @@
+import { Badge } from 'terra-image/package.json?dev-site-package';
+
+<Badge />
+
+# Terra Image Upgrade Guide
+
+## Changes from 2.0 to 3.0
+
+### Changes to CSS Custom Properties
+
+The following transitive dependency had breaking changes. See its upgrade guide for more information.
+* terra-doc-template
