+++ packages/terra-core-docs/src/terra-dev-site/doc/image/example/common/A11yExamples.module.scss
@@ -1,0 +1,73 @@
+:local {
+    .card {
+      width: 100%;
+    }
+  
+    .informative-example {
+      font-size: 0.85rem;
+      margin-left: auto;
+      margin-right: auto;
+      max-width: 40rem;
+      padding: 0.25rem 1rem 0;
+      text-align: center;
+      width: 100%;
+    }
+  
+    .image {
+      margin: 0 auto;
+    }
+  
+    .title {
+      display: block;
+      font-size: 1.1em;
+      margin-bottom: 0.25em;
+      margin-top: 0.5em;
+    }
+  
+    .cite {
+      display: block;
+      margin-top: 0.25em;
+    }
+  
+    .decorative-example {
+      min-height: 240px;
+      position: relative;
+      width: 100%;
+  
+      .background {
+        height: 240px;
+        width: 100%;
+      }
+  
+      .foreground {
+        padding: 20px;
+        position: absolute;
+        top: 0;
+        width: 100%;
+      }
+  
+      .card {
+        opacity: 0.85;
+      }
+    }
+  
+    .decorative-text-container {
+      font-size: 0.85rem;
+      padding-left: 1rem;
+      padding-right: 1rem;
+      text-align: center;
+    }
+  
+    @media screen and (min-width: 544px) {
+      .informative-example {
+        font-size: 1rem;
+      }
+  
+      .decorative-text-container {
+        font-size: 1rem;
+        padding-left: 3rem;
+        padding-right: 3rem;
+      }  
+    }
+  }
+
