+++ packages/terra-image/src/Image.jsx
@@ -1,8 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ImageVariant, FitTypes } from './_BaseImage';
-import InformativeImage from './_InformativeImage';
-import DecorativeImage from './_DecorativeImage';
+import objectFitImages from 'object-fit-images'; // Added polyfill for IE.
+import classNames from 'classnames';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
+import styles from './Image.module.scss';
+import {
+  getDecorativeImage, getInformativeImage, ImageVariant, FitTypes,
+} from './_imageHelper';
+
+const cx = classNamesBind.bind(styles);
 
 const propTypes = {
   /**
@@ -49,15 +56,127 @@ const propTypes = {
   fit: PropTypes.oneOf(['cover', 'scale-down', 'fill', 'contain', 'none']),
 };
 
-const Image = (props) => {
-  if (props.alt !== undefined) {
-    return <InformativeImage {...props} />;
-  }
-
-  return <DecorativeImage {...props} />;
+const defaultProps = {
+  variant: 'default',
+  isFluid: false,
+  fit: 'fill',
 };
 
+class Image extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.ImageRef = React.createRef();
+
+    this.state = {
+      isLoading: true,
+      isError: false,
+      prevPropsSrc: props.src,
+    };
+
+    this.handleOnLoad = this.handleOnLoad.bind(this);
+    this.handleOnError = this.handleOnError.bind(this);
+  }
+
+  componentDidUpdate() {
+    objectFitImages(this.ImageRef.current);
+  }
+
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (nextProps.src !== prevState.prevPropsSrc) {
+      return {
+        isLoading: true,
+        isError: false,
+        prevPropsSrc: nextProps.src,
+      };
+    }
+
+    return null;
+  }
+
+  handleOnLoad() {
+    this.setState({ isLoading: false });
+    const { onLoad } = this.props;
+
+    if (onLoad !== undefined) {
+      onLoad();
+    }
+  }
+
+  handleOnError() {
+    this.setState({ isLoading: false, isError: true });
+    const { onError } = this.props;
+
+    if (onError !== undefined) {
+      onError();
+    }
+  }
+
+  createImage(customProps, imageClasses) {
+    const {
+      src, alt, height, width,
+    } = customProps;
+
+    const imageProps = (alt !== undefined && alt?.trim() !== '') ? getInformativeImage(customProps) : getDecorativeImage(customProps);
+
+    return (
+      <img
+        {...imageProps}
+        src={src}
+        alt={alt}
+        height={height}
+        width={width}
+        onLoad={this.handleOnLoad}
+        onError={this.handleOnError}
+        className={imageClasses}
+        ref={this.ImageRef}
+      />
+    );
+  }
+
+  render() {
+    const {
+      variant, isFluid, placeholder, fit, ...customProps
+    } = this.props;
+
+    const theme = this.context;
+
+    const imageClasses = classNames(
+      cx(
+        'image',
+        fit,
+        variant,
+        theme.className,
+        { fluid: isFluid },
+        { hidden: placeholder && this.state.isLoading },
+      ),
+      customProps.className,
+    );
+
+    delete customProps.className;
+    if (!this.state.isLoading) {
+      objectFitImages(this.ImageRef.current);
+    }
+    if (placeholder) {
+      if (this.state.isLoading) {
+        return (
+          <>
+            {this.createImage(customProps, imageClasses)}
+            {placeholder}
+          </>
+        );
+      }
+
+      return this.state.isError ? placeholder : this.createImage(customProps, imageClasses);
+    }
+
+    return this.createImage(customProps, imageClasses);
+  }
+}
+
 Image.propTypes = propTypes;
+Image.defaultProps = defaultProps;
+Image.contextType = ThemeContext;
 
 export default Image;
 export { ImageVariant, FitTypes };
